Homework 5, task 2: restructurize
