+++ library/numeric/binary_trees/node_operations/basic_operations.h
@@ -1,6 +1,6 @@
 #ifndef SPBU_2020_C_NUMERIC_NODES_BST_OPERATIONS_H
 #define SPBU_2020_C_NUMERIC_NODES_BST_OPERATIONS_H
-#include "../bst_node.h"
+#include "../node_structures/bst_node.h"
 #include <stdbool.h>
 
 int getChildrenCountBinaryTreeNode(BinaryTreeNode*);
