+++ library/numeric/binary_trees/node_structures/avl_node.h
