+++ library/numeric/binary_trees/node_structures/bst_node.h
